// src/store/workspaces/mutation-types.ts
export enum WorkspaceMutationTypes {
  SET_WORKSPACES = 'WORKSPACE/SET_WORKSPACES',
  SET_SHARED_WORKSPACES = 'WORKSPACE/SET_SHARED_WORKSPACES',
  SET_OWNERS_WORKSPACES_LOADING = 'WORKSPACE/SET_OWNERS_WORKSPACES_LOADING',
  SET_OWNERS_WORKSPACES_LOADED = 'WORKSPACE/SET_OWNERS_WORKSPACES_LOADED',
  SET_SHARED_WORKSPACES_LOADING = 'WORKSPACE/SET_SHARED_WORKSPACES_LOADING',
  SET_SHARED_WORKSPACES_LOADED = 'WORKSPACE/SET_SHARED_WORKSPACES_LOADED',
  SET_WORKSPACE_SHARE_CODE = 'WORKSPACE/SET_WORKSPACE_SHARE_CODE'
}